pyproject.toml
[project]
name = "networth-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "supabase>=2.28.0",
    "yfinance>=1.2.0",
]
